pyproject.toml
[tool.poetry]
name = "newspaper"
version = "0.5.1"
description = "A generic news website project"
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-filter = "^22.1"
django-widget-tweaks = "^1.4.12"
django-allauth = "^0.52.0"
python-dotenv = "^0.21.0"
celery = {extras = ["redis"], version = "^5.2.7"}
django-redis = "^5.2.0"
psycopg2-binary = "^2.9.5"
django-modeltranslation = "^0.18.9"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
django-stubs = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
